+++ WVForcingType.m
@@ -1,5 +1,5 @@
 classdef WVForcingType
     enumeration
-        HydrostaticSpatial, NonhydrostaticSpatial, PVSpatial, QGSpatial, Spectral, PVSpectral, QGSpectral, SpectralAmplitude, PVSpectralAmplitude
+        HydrostaticSpatial, NonhydrostaticSpatial, PVSpatial, QGSpatial, Spectral, PVSpectral, QGSpectral, SpectralAmplitude, PVSpectralAmplitude, BoussinesqSpectral
     end
 end